feat: default main layout margin
