+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "dev": "next dev",
     "start": "npx serve@latest out",
+    "start:product": "next build && next export && npm run start",
     "build": "next build && next export",
     "predeploy": "npm run build",
     "deploy": "touch out/.nojekyll && gh-pages -d out --dotfiles"
