+++ src/containers/MainSection/MainSection.scss
@@ -1,5 +1,4 @@
 @use "/src/sass/abstracts/variables";
-
 @use "/src/sass/layout/breakpoint";
 
 .mainSection {
