+++ src/sass/abstracts/_variables.scss
@@ -1,3 +1,7 @@
 // padding, margin
 $page-padding: 16px;
 $page-padding-mobile: 8px;
+
+// max-width
+$page-max-width-pc: 1024px;
+$page-max-width-mobile: 328px;
